requirements.txt
# ==================================
# Core Python Utilities
# ==================================
numpy==1.26.4         # Numerical operations
pandas==2.2.2         # Data handling & preprocessing
scipy==1.13.1         # Scientific computations

# ==================================
# Machine Learning (Classroom Models)
# ==================================
scikit-learn==1.5.1   # Logistic Regression, Decision Tree, Naïve Bayes, K-Means
joblib>=1.3.0         # Model serialization
datasets>=2.0.0       # Hugging Face datasets

# ==================================
# Natural Language Processing (NLP)
# ==================================
nltk==3.9             # Tokenization, stopwords, stemming
spacy==3.7.4          # Named Entity Recognition (skills, education, companies)

# ==================================
# AI & Embeddings
# ==================================
openai>=1.0.0         # OpenAI API for embeddings
pinecone-client>=3.0.0 # Pinecone vector database
tiktoken>=0.5.0       # OpenAI tokenizer for chunking

# ==================================
# Resume & JD Parsing
# ==================================
PyPDF2==3.0.1         # Extract text from PDF resumes
python-docx==1.1.2    # Extract text from DOCX resumes
docx2txt==0.8         # Alternative text extraction from DOCX

# ==================================
# Advanced Models (External / Improved)
# ==================================
sentence-transformers==3.0.1  # BERT embeddings for JD–Resume matching
transformers==4.44.0          # HuggingFace library for NLP models
torch==2.4.1                  # Backend for BERT (PyTorch)

# ==================================
# Database
# ==================================
sqlite-utils==3.36.0  # Manage SQLite database easily

# ==================================
# API (Optional for Final Integration)
# ==================================
fastapi==0.115.0      # Backend API framework
uvicorn==0.30.6       # Run FastAPI server
flask==3.0.3          # Alternative lightweight API framework

# ==================================
# Frontend Helpers
# ==================================
jinja2==3.1.4         # Template rendering (if UI extended)

# ==================================
# Visualization
# ==================================
matplotlib==3.9.2     # Plotting graphs
seaborn==0.13.2       # Heatmaps, confusion matrices

# ==================================
# Jupyter Notebook (For Review Demos)
# ==================================
jupyter==1.1.1        # Run Jupyter environment
notebook==7.2.1       # Jupyter Notebook web interface
